musiticket/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
Django==5.0.1
django-cors-headers==4.3.1
django-extensions==3.2.3
django-rest-framework==0.1.0
djangorestframework==3.14.0
idna==3.6
psycopg2-binary==2.9.9
pycparser==2.21
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.4.4
urllib3==2.2.0
